better descritpions, move Noise_Map_Difference to Experimental
