+++ wps_scripts/src/main/groovy/org/noise_planet/noisemodelling/wps/Experimental_Matsim/Receivers_From_Activities_Closest.groovy
@@ -25,8 +25,8 @@ import org.slf4j.LoggerFactory
 import java.sql.*
 import groovy.sql.Sql
 
-title = 'Chose Receivers From Matsim Activities'
-description = 'Chose one receiver per Mastim Activity. Created a table '
+title = 'Chose Closest Receivers For Matsim Activities'
+description = 'Chose the closest receiver in a RECEIVERS table for every Mastim Activity in an ACTIVITIES table'
 
 inputs = [
         activitiesTable : [
