+++ wps_scripts/src/main/groovy/org/noise_planet/noisemodelling/wps/Import_and_Export/Osm_Pbf_to_Buildings.groovy
@@ -50,7 +50,7 @@ inputs = [
     pathFile: [
         name: 'Path of the OSM file',
         title: 'Path of the OSM file',
-        description: 'Path of the OSM file including extension. </br> For example : c:/home/area.osm.gz',
+        description: 'Path of the OSM file including extension. </br> For example : c:/home/area.osm.pbf',
         type: String.class
     ],
     targetSRID: [
